+++ RealArtists.ShipHub.Api/Models/ApiComment.cs
@@ -2,10 +2,10 @@
   using System;
 
   public class ApiComment {
-    public int Identifier { get; set; }
+    public long Identifier { get; set; }
     public string Body { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
     public DateTimeOffset UpdatedAt { get; set; }
-    public int UserId { get; set; }
+    public long UserId { get; set; }
   }
 }
