+++ RealArtists.ShipHub.Database/Tables/AccountOrganizations.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[AccountOrganizations] (
-  [UserId]         INT NOT NULL,
-  [OrganizationId] INT NOT NULL,
+  [UserId]         BIGINT NOT NULL,
+  [OrganizationId] BIGINT NOT NULL,
   CONSTRAINT [PK_AccountOrganizations] PRIMARY KEY CLUSTERED ([UserId], [OrganizationId]),
   CONSTRAINT [FK_AccountOrganizations_UserId_Accounts_Id] FOREIGN KEY ([UserId]) REFERENCES [dbo].[Accounts] ([Id]),
   CONSTRAINT [FK_AccountOrganizations_OrganizationId_Accounts_Id] FOREIGN KEY ([OrganizationId]) REFERENCES [dbo].[Accounts] ([Id]),
