+++ RealArtists.ShipHub.Database/Tables/RepositoryAccounts.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[RepositoryAccounts] (
-  [RepositoryId] INT NOT NULL,
-  [AccountId]    INT NOT NULL,
+  [RepositoryId] BIGINT NOT NULL,
+  [AccountId]    BIGINT NOT NULL,
   CONSTRAINT [PK_RepositoryAccounts] PRIMARY KEY CLUSTERED ([RepositoryId], [AccountId]),
   CONSTRAINT [FK_RepositoryAccounts_AccountId_Accounts_Id] FOREIGN KEY ([AccountId]) REFERENCES [dbo].[Accounts] ([Id]),
   CONSTRAINT [FK_RepositoryAccounts_RepositoryId_Repositories_Id] FOREIGN KEY ([RepositoryId]) REFERENCES [dbo].[Repositories] ([Id]),
