+++ RealArtists.ShipHub.Database/User Defined Types/MilestoneTableType.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TYPE [dbo].[MilestoneTableType] AS TABLE (
-  [Id]             INT            NOT NULL PRIMARY KEY CLUSTERED,
+  [Id]             BIGINT         NOT NULL PRIMARY KEY CLUSTERED,
   [Number]         INT            NOT NULL,
   [State]          NVARCHAR(10)   NOT NULL,
   [Title]          NVARCHAR(255)  NOT NULL,
